ci: restore gated delta net test grouping
